fix: error_logger tokio::spawn safety + BIGINT for processed_updates
- Use Handle::try_current() instead of bare tokio::spawn to avoid panic
  if ErrorLogger::log() is ever called outside tokio runtime
- Use BIGINT for bot_id/update_id in processed_updates (matches i64 in app code)

+++ migrations/V39__multi_instance_queue.sql
@@ -23,8 +23,8 @@ CREATE UNIQUE INDEX IF NOT EXISTS idx_task_queue_active_idempotency
       AND status IN ('pending', 'leased', 'processing', 'uploading');
 
 CREATE TABLE IF NOT EXISTS processed_updates (
-    bot_id INTEGER NOT NULL,
-    update_id INTEGER NOT NULL,
+    bot_id BIGINT NOT NULL,
+    update_id BIGINT NOT NULL,
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (bot_id, update_id)
 );
